runnable seminar3
