+++ practice/practise-3/build.gradle.kts
@@ -3,6 +3,7 @@ plugins {
 	checkstyle
 	id("org.springframework.boot") version "3.4.2"
 	id("io.spring.dependency-management") version "1.1.7"
+	application
 }
 
 group = "hse"
@@ -42,4 +43,8 @@ dependencies {
 
 tasks.withType<Test> {
 	useJUnitPlatform()
+}
+
+application {
+    mainClass.set("hse.kpo.KpoApplication")
 }
