+++ practice/practise-3/src/main/java/hse/kpo/domains/Customer.java
@@ -7,11 +7,15 @@ import lombok.ToString;
 @Getter
 @ToString
 public class Customer {
+    
+    @Getter
     private final String name;
 
     private final int legPower;
 
     private final int handPower;
+
+    private final int IQ;
 
     @Setter
     private Car car;
@@ -20,5 +24,13 @@ public class Customer {
         this.name = name;
         this.legPower = legPower;
         this.handPower = handPower;
+        this.IQ = 0;
+    }
+
+    public Customer(String name, int legPower, int handPower, int IQ) {
+        this.name = name;
+        this.legPower = legPower;
+        this.handPower = handPower;
+        this.IQ = IQ;
     }
 }
