+++ practice/practise-3/src/main/java/hse/kpo/domains/HandEngine.java
@@ -1,10 +1,17 @@
 package hse.kpo.domains;
 
-import hse.kpo.interfaces.IEngine;
 import lombok.ToString;
+import hse.kpo.interfaces.IEngine;
 
 @ToString
 public class HandEngine implements IEngine {
+    
+    /** 
+     * Проверка совместимости ручного двигателя и покупателя.
+     * 
+     * @param customer - покупатель, с которым мы сравниваем двигатель
+     * @return true, если двигатель подходит покупателю
+     */
     @Override
     public boolean isCompatible(Customer customer) {
         return customer.getHandPower() > 5;
