+++ practice/practise-3/src/main/java/hse/kpo/domains/PedalEngine.java
@@ -1,14 +1,20 @@
 package hse.kpo.domains;
 
-import hse.kpo.interfaces.IEngine;
 import lombok.Getter;
 import lombok.ToString;
+import hse.kpo.interfaces.IEngine;
 
 @ToString
 @Getter
 public class PedalEngine implements IEngine {
     private final int size;
 
+    /** 
+     * Проверка совместимости педального двигателя и покупателя.
+     * 
+     * @param customer - покупатель, с которым мы сравниваем двигатель
+     * @return true, если двигатель подходит покупателю
+     */
     @Override
     public boolean isCompatible(Customer customer) {
         return customer.getLegPower() > 5;
