+++ practice/practise-3/src/main/java/hse/kpo/factories/HandCarFactory.java
@@ -8,6 +8,14 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class HandCarFactory implements ICarFactory<EmptyEngineParams> {
+
+    /**
+     * Метод автомобильной фабрики по сборке ручных автомобилей для создания автомобиля.
+     * 
+     * @param carParams - параметры создаваемого автомобиля EmptyEngineParams
+     * @param carNumber - номер автомобиля
+     * @return - созданный автомобиль
+     */
     @Override
     public Car createCar(EmptyEngineParams carParams, int carNumber) {
         var engine = new HandEngine(); // Создаем двигатель без каких-либо параметров
