+++ practice/practise-3/src/main/java/hse/kpo/factories/PedalCarFactory.java
@@ -1,6 +1,5 @@
 package hse.kpo.factories;
 
-
 import hse.kpo.domains.Car;
 import hse.kpo.domains.PedalEngine;
 import hse.kpo.interfaces.ICarFactory;
@@ -9,6 +8,14 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class PedalCarFactory implements ICarFactory<PedalEngineParams> {
+
+    /**
+     * Метод автомобильной фабрики по сборке педальных автомобилей для создания автомобиля.
+     * 
+     * @param carParams - параметры создаваемого автомобиля PedalEngineParams
+     * @param carNumber - номер автомобиля
+     * @return - созданный автомобиль
+     */
     @Override
     public Car createCar(PedalEngineParams carParams, int carNumber) {
         var engine = new PedalEngine(carParams.pedalSize()); // создаем двигатель на основе переданных параметров
