+++ practice/practise-3/src/main/java/hse/kpo/interfaces/ICarProvider.java
@@ -5,5 +5,11 @@ import hse.kpo.domains.Customer;
 
 public interface ICarProvider {
 
+    /**
+     * Метод для создания получения подходящего для покупателя автомобиля.
+     * 
+     * @param customer - покупатель
+     * @return - подходящий автомобиль
+     */
     Car takeCar(Customer customer); // Метод возвращает optional на Car, что означает, что метод может ничего не вернуть
 }
