+++ practice/practise-3/src/main/java/hse/kpo/interfaces/ICustomerProvider.java
@@ -1,8 +1,15 @@
 package hse.kpo.interfaces;
 
 import hse.kpo.domains.Customer;
+
 import java.util.List;
 
 public interface ICustomerProvider {
+
+    /**
+     * Метод для получения всех покупателей.
+     * 
+     * @return - список покупателей
+     */
     List<Customer> getCustomers(); // метод возвращает коллекцию только для чтения, так как мы не хотим давать вызывающему коду возможность изменять список
 }
