+++ practice/practise-3/src/main/java/hse/kpo/interfaces/IEngine.java
@@ -1,6 +1,5 @@
 package hse.kpo.interfaces;
 
-
 import hse.kpo.domains.Customer;
 
 public interface IEngine {
