+++ practice/practise-3/src/main/java/hse/kpo/services/CustomerStorage.java
@@ -11,11 +11,21 @@ import java.util.List;
 public class CustomerStorage implements ICustomerProvider {
     private final List<Customer> customers = new ArrayList<>();
 
+    /**
+     * Метод для получения всех покупателей из списка customers.
+     * 
+     * @return - список покупателей
+     */
     @Override
     public List<Customer> getCustomers() {
         return customers;
     }
 
+    /**
+     * Метод для добавления покупателя.
+     * 
+     * @param customer - покупатель.
+     */
     public void addCustomer(Customer customer)
     {
         customers.add(customer); // просто добавляем покупателя в список
