+++ practice/practise-3/src/main/java/hse/kpo/services/HseCarService.java
@@ -15,6 +15,22 @@ public class HseCarService {
 
     private final ICustomerProvider customerProvider;
 
+    public HseCarService(ICarProvider carProvider, ICustomerProvider customersProvider)
+    {
+        this.carProvider = carProvider;
+        this.customerProvider = customersProvider;
+    }
+
+    /**
+     * Метод для продажи всех автомобилей всем покупателям.
+     * 
+     * Пробегаемся по всем покупателям, которые не имеют автомобиля.
+     * 
+     * Если у покупателя нет автомобиля, продаем ему подходящий.
+     * 
+     * @see ICarProvider
+     * @see ICustomerProvider
+     */
     public void sellCars()
     {
         // получаем список покупателей
