--- practice/practise-3/src/main/java/studying/domains/Customer.java
@@ -1,36 +1,0 @@
-package studying.domains;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@ToString
-public class Customer {
-    
-    @Getter
-    private final String name;
-
-    private final int legPower;
-
-    private final int handPower;
-
-    private final int IQ;
-
-    @Setter
-    private Car car;
-
-    public Customer(String name, int legPower, int handPower) {
-        this.name = name;
-        this.legPower = legPower;
-        this.handPower = handPower;
-        this.IQ = 0;
-    }
-
-    public Customer(String name, int legPower, int handPower, int IQ) {
-        this.name = name;
-        this.legPower = legPower;
-        this.handPower = handPower;
-        this.IQ = IQ;
-    }
-}
