--- practice/practise-3/src/main/java/studying/interfaces/IEngine.java
@@ -1,14 +1,0 @@
-package studying.interfaces;
-
-import studying.domains.Customer;
-
-public interface IEngine {
-
-    /**
-     * Метод для проверки совместимости двигателя с покупателем.
-     *
-     * @param customer - покупатель, с которым мы сравниваем двигатель
-     * @return true, если двигатель подходит покупателю
-     */
-    boolean isCompatible(Customer customer);
-}
