--- practice/practise-3/src/main/java/studying/params/EmptyEngineParams.java
@@ -1,5 +1,0 @@
-package studying.params;
-
-public record EmptyEngineParams() {
-    public static final EmptyEngineParams DEFAULT = new EmptyEngineParams();
-}
