--- practice/practise-3/src/main/java/studying/params/PedalEngineParams.java
@@ -1,4 +1,0 @@
-package studying.params;
-
-public record PedalEngineParams(int pedalSize) {
-}
